feat: 여행지 목록 조회시 좋아요 개수 반환값에 추가

+++ src/main/java/com/ssafy/trip/domain/like/service/LikeService.java
@@ -4,7 +4,7 @@ public interface LikeService {
 
 //    void addLike(LikeDto likeDto) throws Exception;
 //    void deleteLike(Long likeId) throws Exception;
-    int count(Long spotId) throws Exception;
+    int count(int spotId) throws Exception;
 
     void addOrDeleteLike(Long userId, Long spotId) throws Exception;
